Use double precision arithmetic for conversion

+++ src/main/java/org/esa/cci/sst/rules/AbstractFloatFormat.java
@@ -16,23 +16,23 @@ abstract class AbstractFloatFormat<S extends Number> extends AbstractFormat<S, F
 
     @Override
     protected final Float getTargetFillValue(Number sourceFillValue, Number scaleFactor, Number addOffset) {
-        final float a = getFloat(scaleFactor, 1.0f);
-        final float b = getFloat(addOffset, 0.0f);
-        return apply(sourceFillValue.floatValue(), a, b);
+        final double a = getDouble(scaleFactor, 1.0);
+        final double b = getDouble(addOffset, 0.0);
+        return apply(sourceFillValue.doubleValue(), a, b);
     }
 
     @Override
     protected final void apply(Array sourceArray, Array targetArray, Number scaleFactor, Number addOffset) {
-        final float a = getFloat(scaleFactor, 0.0f);
-        final float b = getFloat(addOffset, 0.0f);
+        final double a = getDouble(scaleFactor, 0.0);
+        final double b = getDouble(addOffset, 0.0);
         for (int i = 0; i < sourceArray.getSize(); i++) {
-            targetArray.setFloat(i, apply(sourceArray.getFloat(i), a, b));
+            targetArray.setFloat(i, apply(sourceArray.getDouble(i), a, b));
         }
     }
 
-    protected abstract float apply(float number, float scaleFactor, float addOffset);
+    protected abstract float apply(double number, double scaleFactor, double addOffset);
 
-    private float getFloat(Number number, float defaultValue) {
+    private double getDouble(Number number, double defaultValue) {
         if (number == null) {
             return defaultValue;
         }
